trading_bot/requirements.txt
python-dotenv==1.0.0
ccxt==4.2.0
pandas==2.1.4
numpy==1.26.3
ta==0.11.0
websocket-client==1.7.0
websockets==12.0
streamlit==1.29.0
plotly==5.18.0
pytest==7.4.3
pytest-asyncio==0.23.2
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3